move some files into leader-schedule
