+++ ledger/src/lib.rs
@@ -22,7 +22,7 @@ pub mod blockstore_processor;
 pub mod entry_notifier_interface;
 pub mod entry_notifier_service;
 pub mod genesis_utils;
-pub mod leader_schedule;
+pub use solana_leader_schedule as leader_schedule;
 pub mod leader_schedule_cache;
 pub mod leader_schedule_utils;
 pub mod next_slots_iterator;
